Make sure we also look at user orgs for sponsorships
Organizations themselves cannot create issues and pull requests,
so we also need to check for sponsoring organizations and whether
the sender belongs to one of them. We take the higher-tier org
and use that to apply the label.

Review down the road if this is too broad of a criteria, since
large orgs like dotnet, for example, often have many users that
aren't necessarily direct employees of the parent company (Microsoft
in that case, say).